feat: rename app to Minimalist
- Change app name from AppManager to Minimalist
- Update applicationId to atsumi.android.minimalist
- Update namespace to atsumi.android.minimalist
- Update all package declarations in Kotlin files
- Update import statements across all files
