+++ app/src/main/java/atsumi/android/appmanager/util/DisplayCondition.kt
@@ -1,4 +1,4 @@
-package atsumi.android.appmanager.util
+package atsumi.android.minimalist.util
 
 interface DisplayCondition<T> {
     fun isDisplayable(obj: T): Boolean
